Add Initial RealEstate Services & Repos, Some Refactoring
