--- src/main/java/ir/ac/kntu/divar/model/entity/Console.java
@@ -1,16 +1,0 @@
-package ir.ac.kntu.divar.model.entity;
-
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.Table;
-
-@Data
-@NoArgsConstructor
-@Table(name = "CONSOLE")
-@Entity(name = "Console")
-@EqualsAndHashCode(callSuper = true)
-public class Console extends Computer {
-}
