+++ src/main/java/ir/ac/kntu/divar/model/entity/advertisement/electronics/Computer.java
@@ -1,4 +1,4 @@
-package ir.ac.kntu.divar.model.entity;
+package ir.ac.kntu.divar.model.entity.advertisement.electronics;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
