+++ src/main/java/ir/ac/kntu/divar/model/entity/advertisement/electronics/ElectronicsAdvertisement.java
@@ -1,0 +1,21 @@
+package ir.ac.kntu.divar.model.entity.advertisement.electronics;
+
+import ir.ac.kntu.divar.model.entity.advertisement.Advertisement;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Entity;
+import javax.persistence.Inheritance;
+import javax.persistence.InheritanceType;
+import javax.persistence.Table;
+
+@Data
+@NoArgsConstructor
+@Table(name = "ELECTRONICS")
+@Entity(name = "Electronics")
+@EqualsAndHashCode(callSuper = true)
+@Inheritance(strategy = InheritanceType.JOINED)
+public abstract class ElectronicsAdvertisement extends Advertisement {
+    private Double price;
+}
