+++ src/main/java/ir/ac/kntu/divar/model/entity/advertisement/realestate/ResidentialRent.java
@@ -1,4 +1,4 @@
-package ir.ac.kntu.divar.model.entity;
+package ir.ac.kntu.divar.model.entity.advertisement.realestate;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
