+++ src/main/java/ir/ac/kntu/divar/model/entity/location/City.java
@@ -1,5 +1,6 @@
-package ir.ac.kntu.divar.model.entity;
+package ir.ac.kntu.divar.model.entity.location;
 
+import ir.ac.kntu.divar.model.entity.BaseEntity;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
