+++ src/main/java/ir/ac/kntu/divar/model/entity/location/Zone.java
@@ -1,5 +1,6 @@
-package ir.ac.kntu.divar.model.entity;
+package ir.ac.kntu.divar.model.entity.location;
 
+import ir.ac.kntu.divar.model.entity.BaseEntity;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
