+++ src/main/java/ir/ac/kntu/divar/model/repo/location/CityRepository.java
@@ -1,6 +1,6 @@
-package ir.ac.kntu.divar.model.repo;
+package ir.ac.kntu.divar.model.repo.location;
 
-import ir.ac.kntu.divar.model.entity.City;
+import ir.ac.kntu.divar.model.entity.location.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
