+++ src/main/java/ir/ac/kntu/divar/model/repo/location/ZoneRepository.java
@@ -1,7 +1,7 @@
-package ir.ac.kntu.divar.model.repo;
+package ir.ac.kntu.divar.model.repo.location;
 
-import ir.ac.kntu.divar.model.entity.City;
-import ir.ac.kntu.divar.model.entity.Zone;
+import ir.ac.kntu.divar.model.entity.location.City;
+import ir.ac.kntu.divar.model.entity.location.Zone;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
