+++ src/main/resources/templates/index.html
@@ -6,7 +6,6 @@
           content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=0">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-    <link rel="manifest" href="/manifest.json">
     <link rel="apple-touch-icon" sizes="180x180"
           href="https://s100.divarcdn.com/static/thewall-assets/apple-touch-icon.png">
     <link rel="icon" type="image/png" sizes="32x32"
